uci_seq_num size correction in SESSION_DATA_TRANSFER_STATUS_NTF
Bug: 291177310
Test: Test data transfer

+++ service/uci/jni/src/notification_manager_android.rs
@@ -959,7 +959,7 @@ impl NotificationManagerAndroid {
     fn on_data_transfer_status_notification(
         &mut self,
         session_id: u32,
-        uci_sequence_number: u8,
+        uci_sequence_number: u16,
         status_code: u8,
         tx_count: u8,
     ) -> Result<JObject, JNIError> {
